src/requirements.txt
altair==4.1.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto3==1.14.2
botocore==1.17.2
Bottleneck==1.3.2
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cmake==3.17.3
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fastai==1.0.61
fastprogress==0.2.3
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pydeck==0.4.0b2
Pygments==2.6.1
pynvx==1.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorchcv==0.0.58
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
spacy==2.2.4
srsly==1.0.2
streamlit==0.61.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
toml==0.10.1
toolz==0.10.0
torch==1.5.0
torchvision==0.6.0
tornado==5.1.1
tqdm==4.46.1
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
wasabi==0.6.0
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
